+++ lib/Data/Time/LocalTime/Internal/TimeOfDay.hs
@@ -5,6 +5,7 @@ module Data.Time.LocalTime.Internal.TimeOfDay
 (
     -- * Time of day
     TimeOfDay(..),midnight,midday,makeTimeOfDayValid,
+    timeToDaysAndTimeOfDay,daysAndTimeOfDayToTime,
     utcToLocalTimeOfDay,localToUTCTimeOfDay,
     timeToTimeOfDay,timeOfDayToTime,
     dayFractionToTimeOfDay,timeOfDayToDayFraction
@@ -17,6 +18,7 @@ import Data.Fixed
 import Data.Data
 #endif
 import Data.Time.Clock.Internal.DiffTime
+import Data.Time.Clock.Internal.NominalDiffTime
 import Data.Time.Calendar.Private
 import Data.Time.LocalTime.Internal.TimeZone
 
@@ -60,6 +62,20 @@ makeTimeOfDayValid h m s = do
     _ <- clipValid 0 59 m
     _ <- clipValid 0 60.999999999999 s
     return (TimeOfDay h m s)
+
+-- | Convert a period of time into a count of days and a time of day since midnight.
+-- The time of day will never have a leap second.
+timeToDaysAndTimeOfDay :: NominalDiffTime -> (Integer,TimeOfDay)
+timeToDaysAndTimeOfDay dt = let
+    s = realToFrac dt
+    (m,ms) = divMod' s 60
+    (h,hm) = divMod' m 60
+    (d,dh) = divMod' h 24
+    in (d,TimeOfDay dh hm ms)
+
+-- | Convert a count of days and a time of day since midnight into a period of time.
+daysAndTimeOfDayToTime :: Integer -> TimeOfDay -> NominalDiffTime
+daysAndTimeOfDayToTime d (TimeOfDay dh hm ms) = (+) (realToFrac ms) $ (*) 60 $ (+) (realToFrac hm) $ (*) 60 $ (+) (realToFrac dh) $ (*) 24 $ realToFrac d
 
 -- | Convert a time of day in UTC to a time of day in some timezone, together with a day adjustment.
 utcToLocalTimeOfDay :: TimeZone -> TimeOfDay -> (Integer,TimeOfDay)