+++ test/main/Main.hs
@@ -17,6 +17,7 @@ import Test.Clock.TAI
 import Test.Format.Format
 import Test.Format.ParseTime
 import Test.LocalTime.Time
+import Test.LocalTime.TimeOfDay
 
 
 tests :: TestTree
@@ -43,7 +44,8 @@ tests = testGroup "Time" [
         testParseTime
         ],
     testGroup "LocalTime" [
-        testTime
+        testTime,
+        testTimeOfDay
         ]
     ]